// src/app/models/book.model.ts
export interface IBook {
  id?: string | undefined;
  title: string | undefined;
  author: string | undefined;
  genre: string | undefined;
  description: string | undefined;
  year: number | undefined;
  fileName: string | undefined;
  imgURL: string | undefined;
  userID: string | undefined;
  date: number | undefined;
}